requirements.txt
mysql-connector-python==8.0.32
Flask==1.24.2
faker==17.0.0
numpy==1.24.2